% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_p_computation.R
\name{generate_distribution_xi_t}
\alias{generate_distribution_xi_t}
\title{Compute the xi for all the iterations in each image
This function takes list of output images, mask files, a null file, total number of iterations and the thinning factor
 For each output image and for each mask file, it computes a distribution of Xi
 Each output image is read once and computations are made on it}
\usage{
generate_distribution_xi_t(images, mask_files, null_file, n_iter)
}
\description{
Compute the xi for all the iterations in each image
This function takes list of output images, mask files, a null file, total number of iterations and the thinning factor
 For each output image and for each mask file, it computes a distribution of Xi
 Each output image is read once and computations are made on it
}
